Extract cache into own file
